metric/requirements.txt
customtkinter
pillow
matplotlib
openpyxl
fpdf
google-generativeai
